Make changes in the script getWikiLinks_with_logging.py
The script getWikiLinks_with_logging.py include many small changes
that I decided to make to it for various reasons. The script can
currently saves the end of the url, the title of the website and
the beginning of the page content, but so far in an "unreadable"
format with tags.