Make changes in the script getWikiLinks_with_logging.py
The script getWikiLinks_with_logging.py work now so that it prints
and saves the end of Wikipedia url, the title of the page, and
the beginning of the page body. The individual contents related in
this way are separated by a separator. This will allow to see what
is on the found page.